Apply library mode to other formats and document them
