+++ runtime/commonMain/src/kotlinx/serialization/internal/Util.kt
@@ -9,19 +9,12 @@ import kotlinx.serialization.*
 
 @InternalSerializationApi
 @Deprecated(
-    level = DeprecationLevel.WARNING,
+    level = DeprecationLevel.ERROR,
     message = "readExactNBytes slipped into public API surface accidentally and will be removed in the future releases. " +
             "You can copy-paste it to your project or (better) find a polished implementation that initially was intended for public uses."
 )
 public fun InputStream.readExactNBytes(bytes: Int): ByteArray {
-    val array = ByteArray(bytes)
-    var read = 0
-    while (read < bytes) {
-        val i = this.read(array, read, bytes - read)
-        if (i == -1) error("Unexpected EOF")
-        read += i
-    }
-    return array
+    throw NotImplementedError("Deprecated")
 }
 
 internal object InternalHexConverter {
@@ -72,7 +65,7 @@ internal object InternalHexConverter {
 
 @InternalSerializationApi
 @Deprecated(
-    level = DeprecationLevel.WARNING,
+    level = DeprecationLevel.ERROR,
     message = "HexConverter slipped into public API surface accidentally and will be removed in the future releases. " +
             "You can copy-paste it to your project or (better) find a polished implementation that initially was intended for public uses."
 )
